Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/content-security-policy/font-src/font-stylesheet-font-blocked.sub-expected.txt


FAIL Test font does not load if it does not match font-src. assert_equals: expected "font-src" but got "font-src 'none'"
